+++ apps/backend/src/libs/modules/database/libs/enums/database-table-name.enum.ts
@@ -9,6 +9,7 @@ const DatabaseTableName = {
 	FILES: "files",
 	FRIENDS: "friends",
 	MIGRATIONS: "migrations",
+	NOTIFICATIONS: "notifications",
 	SECTION_STATUSES: "section_statuses",
 	USER_DETAILS: "user_details",
 	USERS: "users",