+++ apps/backend/src/modules/friends/friends.ts
@@ -1,4 +1,5 @@
 import { logger } from "~/libs/modules/logger/logger.js";
+import { notificationService } from "~/modules/notifications/notifications.js";
 import { UserModel } from "~/modules/users/user.model.js";
 
 import { FriendController } from "./friend.controller.js";
@@ -6,7 +7,7 @@ import { FriendRepository } from "./friend.repository.js";
 import { FriendService } from "./friend.service.js";
 
 const friendRepository = new FriendRepository(UserModel);
-const friendService = new FriendService(friendRepository);
+const friendService = new FriendService(friendRepository, notificationService);
 const friendController = new FriendController(logger, friendService);
 
 export { friendController };