+++ apps/frontend/src/libs/enums/app-route.enum.ts
@@ -6,6 +6,7 @@ const AppRoute = {
 	FRIENDS: "/friends",
 	FRIENDS_FOLLOWERS: "/friends/followers",
 	FRIENDS_FOLLOWINGS: "/friends/followings",
+	NOTIFICATIONS: "/notifications",
 	PROFILE: "/profile",
 	ROOT: "/",
 	SIGN_IN: "/sign-in",