+++ apps/frontend/src/libs/helpers/helpers.ts
@@ -7,5 +7,6 @@ export {
 	getFormattedDate,
 	getPercentage,
 	getSanitizedHtml,
+	getTimeDistanceFormatDate,
 	initDebounce,
 } from "@trackmates/shared";