+++ apps/frontend/src/modules/chats/libs/constants/default-search-chat-payload.ts
@@ -1,0 +1,7 @@
+import { type ChatSearchResponseDto } from "../types/types.js";
+
+const DEFAULT_SEARCH_CHAT_PAYLOAD: ChatSearchResponseDto = {
+	search: "",
+};
+
+export { DEFAULT_SEARCH_CHAT_PAYLOAD };