+++ packages/shared/src/index.ts
@@ -21,6 +21,7 @@ export {
 	getPercentage,
 	getSanitizedHtml,
 	getSizeInBytes,
+	getTimeDistanceFormatDate,
 	initDebounce,
 } from "./libs/helpers/helpers.js";
 export { type Config } from "./libs/modules/config/config.js";
@@ -80,6 +81,7 @@ export {
 	type ChatGetAllItemResponseDto,
 	type ChatItemResponseDto,
 	type ChatResponseDto,
+	type ChatSearchResponseDto,
 	ChatsApiPath,
 	chatCreateValidationSchema,
 	chatIdParameterValidationSchema,
@@ -124,6 +126,15 @@ export {
 	friendIdParameterValidationSchema,
 } from "./modules/friends/friends.js";
 export {
+	type AllNotificationsResponseDto,
+	type CreateNotificationRequestDto,
+	NotificationError,
+	type NotificationResponseDto,
+	NotificationStatus,
+	NotificationType,
+	type UpdateNotificationRequestDto,
+} from "./modules/notifications/notifications.js";
+export {
 	SectionStatus,
 	type SectionStatusAddRequestDto,
 	SectionStatusError,
@@ -143,6 +154,11 @@ export {
 	userCourseGetAllQueryValidationSchema,
 	userIdParameterValidationSchema,
 } from "./modules/user-courses/users-courses.js";
+export {
+	type ReadNotificationsRequestDto,
+	UserNotificationsApiPath,
+	readNotificationsRequestValidationSchema,
+} from "./modules/user-notifications/user-notifications.js";
 export {
 	type UserAuthResponseDto,
 	UserError,