+++ packages/shared/src/modules/user-notifications/types/read-notifications-reaquest-dto.type.ts
@@ -1,0 +1,5 @@
+type ReadNotificationsRequestDto = {
+	notificationIds: number[];
+};
+
+export { type ReadNotificationsRequestDto };